move ga4gh data models back into ga4gh.py to top on request from Kyle
