+++ fence/blueprints/ga4gh.py
@@ -20,9 +20,6 @@ from fence.auth import login_required
 from fence.errors import UserError
 from fence.config import config
 
-from fence.resources.ga4gh.ga4gh_data_models import (
-    BulkObjectAccessRequest,
-)
 
 from fence.blueprints.data.indexd import (
     get_signed_url_for_file,
@@ -31,6 +28,49 @@ from fence.blueprints.data.indexd import (
 )
 from fence.models import UserPassport
 from fence.resources.user import get_current_user_info
+
+
+class BulkObjectAccessIds(BaseModel):
+    bulk_object_id: str
+    bulk_access_ids: list[str]
+
+
+class BulkObjectAccessRequest(BaseModel):
+    passports: Optional[list[str]] = None
+    bulk_object_access_ids: list[BulkObjectAccessIds]
+
+    def map_access_to_object_ids(self):
+        result = defaultdict(list)
+        for item in self.bulk_object_access_ids:
+            for access_id in item.bulk_access_ids:
+                result[access_id].append(item.bulk_object_id)
+
+        return result
+
+
+class ResolvedDrsObject(BaseModel):
+    drs_object_id: str
+    drs_access_id: str
+    url: str
+    headers: Optional[str]
+
+
+class UnresolvedDrsObject(BaseModel):
+    error_code: int
+    object_ids: list[str]
+
+
+class BulkObjectSummary(BaseModel):
+    requested: int
+    resolved: int
+    unresolved: int
+
+
+class BulkObjectAccessResponse(BaseModel):
+    summary: BulkObjectSummary
+    unresolved_drs_objects: UnresolvedDrsObject
+    resolved_drs_object_access_urls: ResolvedDrsObject
+
 
 blueprint = Blueprint("ga4gh", __name__)
 logger = get_logger(__name__)
